+++ src/whatsapp/dto/instance.dto.ts
@@ -4,6 +4,7 @@ export class InstanceDto {
   webhook_by_events?: boolean;
   events?: string[];
   qrcode?: boolean;
+  number?: string;
   token?: string;
   chatwoot_account_id?: string;
   chatwoot_token?: string;