+++ src/whatsapp/types/wa.types.ts
@@ -25,9 +25,15 @@ export enum Events {
 }
 
 export declare namespace wa {
-  export type QrCode = { count?: number; base64?: string; code?: string };
+  export type QrCode = {
+    count?: number;
+    pairingCode?: string;
+    base64?: string;
+    code?: string;
+  };
   export type Instance = {
     qrcode?: QrCode;
+    pairingCode?: string;
     authState?: { state: AuthenticationState; saveCreds: () => void };
     name?: string;
     wuid?: string;
@@ -49,6 +55,12 @@ export declare namespace wa {
     url?: string;
     name_inbox?: string;
     sign_msg?: boolean;
+  };
+
+  export type LocalSettings = {
+    reject_call?: boolean;
+    msg_call?: string;
+    groups_ignore?: boolean;
   };
 
   export type StateConnection = {